requirements.txt
SpeechRecognition==3.10.0
pydub==0.25.1
google-cloud-speech==2.21.0
rich==13.4.2
python-dotenv==1.0.0
anthropic==0.3.6